Reorganizes project structure as discussed in discussion #117.
