+++ app/src/main/java/com/musicplayer/SocyMusic/MediaPlayerUtil.java
@@ -6,6 +6,8 @@ import android.net.Uri;
 
 import androidx.annotation.NonNull;
 
+import com.musicplayer.SocyMusic.ui.main.MainActivity;
+
 public class MediaPlayerUtil {
     private static MediaPlayer mediaPlayer;
