+++ app/src/main/res/layout/fragment_dir_browser.xml
@@ -15,7 +15,7 @@
         android:textSize="18sp"
         tools:ignore="HardcodedText" />
 
-    <com.musicplayer.SocyMusic.custom_views.EmptyRecyclerView
+    <com.musicplayer.SocyMusic.custom_views.CustomRecyclerView
         android:id="@+id/recyclerview_dir_browser"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
