+++ app/src/main/res/layout/fragment_player.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/player_background"
-    tools:context="com.musicplayer.SocyMusic.PlayerFragment">
+    tools:context="com.musicplayer.SocyMusic.ui.player.PlayerFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/layout_player_holder"
